move feature size validating tests into a separate suite
